tooltips my beloathed

+++ Content/Projectiles/MartianLightning.cs
@@ -60,7 +60,7 @@ public class MartianLightning : ModProjectile
             Projectile.friendly = true;
             Projectile.ai[0]++;
 
-            if (Projectile.ai[0] > 60 && (!Owner.channel || Projectile.ai[1] != 0))
+            if (Projectile.ai[0] > 60 && !Owner.channel || (Projectile.ai[0] > 20 && Projectile.ai[1] != 0))
                 Projectile.Kill();
 
             if (Projectile.ai[0] == 1)
